requirements.txt
dwave-cloud-client>=0.8.0
numpy>=1.20.0